Django_learning/dataAnalysisPlatform/routeAnalysisPlatform/requirement.txt
django-filer
djangorestframework
django-cors-headers
markdown        # Markdown support for the browsable API.
django-contrib-postgres==0.0.1